Can now add results to an existing file.

+++ scripts/process_scan.py
@@ -4,8 +4,9 @@
 # argv[1]: fasta file containing proteome
 # argv[2]: path to hmmer_scan.sh results
 # argv[3]: level
-# argv[4]: path and name for new text file
+# argv[4]: path and name for new text file or existing file to add to
 
+import os.path
 import sys
 from Bio import SeqIO
 from Bio import SearchIO
@@ -23,14 +24,17 @@ def process_scan():
 		else:
 			OG = protein[0].id.replace(".meta_raw", "")
 		processed.append((sys.argv[3]+" ", protein.id+" ", OG+" ", pid))	
-	output = open(sys.argv[4], "w")
-	output.write("Level ProteinID GroupID ProteomeID \n")
-
+	if os.path.isfile(sys.argv[4]):
+		output = open(sys.argv[4], "a")
+	else:
+		output = open(sys.argv[4], "w")
+		output.write("Level ProteinID GroupID ProteomeID \n")
+		
 	for i in processed:
 		output.write("".join(str(s) for s in i) + "\n")
 	output.close()	
 
-if len(sys.argv) != 5:
+if len(sys.argv) !=5:
 	print("Incorrect inputs, needs proteome.fasta, path/to/hmmer results, level, and output name.")
 else:
 	process_scan()
